New seth (#17)
* admin/ can add user as cust, admin, both

* fixed register user

* admin add and reg work 4/5

* admin add and reg work 1:34

+++ src/java/JavaObjects/ViewCollectionBean.java
@@ -13,6 +13,8 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 import javax.annotation.PostConstruct;
 import javax.annotation.Resource;
+import javax.faces.application.FacesMessage;
+import javax.faces.context.FacesContext;
 import javax.sql.DataSource;
 
 /**
@@ -67,7 +69,7 @@ public class ViewCollectionBean implements Serializable {
 
             while (result.next()) {
                 Album a = new Album();
-                a.setAlbumID(result.getInt("ALBUM_ID"));
+                a.setAlbumID(result.getLong("ALBUM_ID"));
                 a.setTitle(result.getString("TITLE"));
                 a.setArtist(result.getString("ARTIST"));
                 a.setReleaseYear(result.getInt("YEAR"));
